Update espsense-bn-link-example.yaml
Added reboot_timeout 0s to both wifi and API.  I believe it's the API that's causing the random reboots.   The API doesn't actually fail because connections are still accepted...but after about 15 minutes the api will error out.  This causes the reboot timer to start (15 minutes default).  Setting it to 0 disables the reboot timer.  Seems to be a bug in the ESPHome code.  There are lots of issues posted on ESPHome related to this, but no common thread of solution.